alias decorator config moved out from modules root
